application state now stored in localStorage
added new dependency "react-localstorage." glue code is a bit terse, and
not what is shown in the README (which uses mixins, which are deprecated.)
The package is otherwise quite sound IMO, and I think my integration path
is better than the mixin path.

this addresses issue #4, however, pending a "new game" feature, I will keep
the issue open.