remember player names; checkbox to toggle players
player names are now added to the state of the App component, which is
synchronized to localStorage, so they are remembered between tabs and
games.

App component state now has a stateVersion property initialized to
STATE_VERSION. please bump this constant if your app is not backward
compatible with the previous version of the app's state. when the app is
loaded and discovers incompatible state, state is simply reinitialized for
the new version.

Remembered players can be added with checkbox toggle.

Current players can be removed with checkbox toggle.

Remembered players cannot be sorted, though maybe this is a feature worth
adding in the future. The user feedback when you fail to sort because you
tried sorting a remembered player, or because you tried sorting a current
player to place it among the remembered players, is non-existent. This is
bad UX.

The checkbox is also really ugly atm!

New Game function was added to nav to allow reinitializing the game.

added lodash as a dependency

finally actually removed test players

The components powering the Edit Players funtion of the app is starting to
look a little hairy. Maybe time to generalize?

The calculations required to map from App component state to props for
PlayerListEdit are also looking pretty hairy.